+++ cli/src/dag/dag_execute.rs
@@ -23,28 +23,15 @@ pub(crate) async fn execute_dag(
     command_title!("Executing Nexus DAG '{dag_id}'");
 
     // Load CLI configuration.
-    let conf = CliConf::load().await.unwrap_or_else(|_| CliConf::default());
+    let conf = CliConf::load().await.unwrap_or_default();
 
     // Nexus objects must be present in the configuration.
-    let NexusObjects {
-        workflow_pkg_id,
-        primitives_pkg_id,
-        network_id,
-        default_sap,
-        gas_service,
-        ..
-    } = get_nexus_objects(&conf)?;
+    let objects = get_nexus_objects(&conf)?;
 
     // Create wallet context, Sui client and find the active address.
     let mut wallet = create_wallet_context(&conf.sui.wallet_path, conf.sui.net).await?;
     let sui = build_sui_client(&conf.sui).await?;
-
-    let address = match wallet.active_address() {
-        Ok(address) => address,
-        Err(e) => {
-            return Err(NexusCliError::Any(e));
-        }
-    };
+    let address = wallet.active_address().map_err(NexusCliError::Any)?;
 
     // Fetch gas coin object.
     let gas_coin = fetch_gas_coin(&sui, conf.sui.net, address, sui_gas_coin).await?;
@@ -60,24 +47,11 @@ pub(crate) async fn execute_dag(
 
     let mut tx = sui::ProgrammableTransactionBuilder::new();
 
-    match dag::execute(
-        &mut tx,
-        default_sap,
-        &dag,
-        gas_service,
-        &entry_group,
-        input_json,
-        *workflow_pkg_id,
-        *primitives_pkg_id,
-        *network_id,
-    ) {
-        Ok(tx) => tx,
-        Err(e) => {
-            tx_handle.error();
+    if let Err(e) = dag::execute(&mut tx, objects, &dag, &entry_group, input_json) {
+        tx_handle.error();
 
-            return Err(NexusCliError::Any(e));
-        }
-    };
+        return Err(NexusCliError::Any(e));
+    }
 
     tx_handle.success();
 
@@ -102,7 +76,7 @@ pub(crate) async fn execute_dag(
                 object_type,
                 object_id,
                 ..
-            } if object_type.address == **workflow_pkg_id
+            } if object_type.address == *objects.workflow_pkg_id
                 && object_type.module == workflow::Dag::DAG_EXECUTION.module.into()
                 && object_type.name == workflow::Dag::DAG_EXECUTION.name.into() =>
             {