+++ cli/src/dag/dag_publish.rs
@@ -23,25 +23,15 @@ pub(crate) async fn publish_dag(
     command_title!("Publishing Nexus DAG");
 
     // Load CLI configuration.
-    let conf = CliConf::load().await.unwrap_or_else(|_| CliConf::default());
+    let conf = CliConf::load().await.unwrap_or_default();
 
     // Nexus objects must be present in the configuration.
-    let NexusObjects {
-        workflow_pkg_id,
-        primitives_pkg_id,
-        ..
-    } = get_nexus_objects(&conf)?;
+    let objects = get_nexus_objects(&conf)?;
 
     // Create wallet context, Sui client and find the active address.
     let mut wallet = create_wallet_context(&conf.sui.wallet_path, conf.sui.net).await?;
     let sui = build_sui_client(&conf.sui).await?;
-
-    let address = match wallet.active_address() {
-        Ok(address) => address,
-        Err(e) => {
-            return Err(NexusCliError::Any(e));
-        }
-    };
+    let address = wallet.active_address().map_err(NexusCliError::Any)?;
 
     // Fetch gas coin object.
     let gas_coin = fetch_gas_coin(&sui, conf.sui.net, address, sui_gas_coin).await?;
@@ -55,10 +45,10 @@ pub(crate) async fn publish_dag(
     let mut tx = sui::ProgrammableTransactionBuilder::new();
 
     // Create an empty DAG.
-    let mut dag_arg = dag::empty(&mut tx, *workflow_pkg_id);
+    let mut dag_arg = dag::empty(&mut tx, objects);
 
     // Create DAG PTB from Dag struct.
-    dag_arg = match dag::create(&mut tx, *workflow_pkg_id, *primitives_pkg_id, dag_arg, dag) {
+    dag_arg = match dag::create(&mut tx, objects, dag_arg, dag) {
         Ok(dag_arg) => dag_arg,
         Err(e) => {
             tx_handle.error();
@@ -68,7 +58,7 @@ pub(crate) async fn publish_dag(
     };
 
     // Public share the DAG, locking it.
-    dag::publish(&mut tx, *workflow_pkg_id, dag_arg);
+    dag::publish(&mut tx, objects, dag_arg);
 
     tx_handle.success();
 
@@ -93,7 +83,7 @@ pub(crate) async fn publish_dag(
                 object_type,
                 object_id,
                 ..
-            } if object_type.address == **workflow_pkg_id
+            } if object_type.address == *objects.workflow_pkg_id
                 && object_type.module == workflow::Dag::DAG.module.into()
                 && object_type.name == workflow::Dag::DAG.name.into() =>
             {