+++ cli/src/gas/gas_add_budget.rs
@@ -12,33 +12,23 @@ pub(crate) async fn add_gas_budget(
     command_title!("Adding '{coin}' as gas budget for Nexus");
 
     // Load CLI configuration.
-    let conf = CliConf::load().await.unwrap_or_else(|_| CliConf::default());
+    let conf = CliConf::load().await.unwrap_or_default();
 
     // Nexus objects must be present in the configuration.
-    let NexusObjects {
-        workflow_pkg_id,
-        gas_service,
-        ..
-    } = get_nexus_objects(&conf)?;
+    let objects = get_nexus_objects(&conf)?;
 
     // Create wallet context, Sui client and find the active address.
     let mut wallet = create_wallet_context(&conf.sui.wallet_path, conf.sui.net).await?;
     let sui = build_sui_client(&conf.sui).await?;
-
-    let address = match wallet.active_address() {
-        Ok(address) => address,
-        Err(e) => {
-            return Err(NexusCliError::Any(e));
-        }
-    };
+    let address = wallet.active_address().map_err(NexusCliError::Any)?;
 
     // Fetch gas coin object.
     let gas_coin = fetch_gas_coin(&sui, conf.sui.net, address, sui_gas_coin).await?;
 
     // Fetch budget coin.
-    let coin = fetch_object_by_id(&sui, coin).await?;
+    let budget_coin = fetch_object_by_id(&sui, coin).await?;
 
-    if coin.object_id == gas_coin.coin_object_id {
+    if budget_coin.object_id == gas_coin.coin_object_id {
         return Err(NexusCliError::Any(anyhow!(
             "Gas and budget coins must be different."
         )));
@@ -47,25 +37,16 @@ pub(crate) async fn add_gas_budget(
     // Fetch reference gas price.
     let reference_gas_price = fetch_reference_gas_price(&sui).await?;
 
-    // Craft a TX to publish the DAG.
+    // Craft the transaction.
     let tx_handle = loading!("Crafting transaction...");
 
     let mut tx = sui::ProgrammableTransactionBuilder::new();
 
-    match gas::add_budget(
-        &mut tx,
-        *workflow_pkg_id,
-        gas_service,
-        address.into(),
-        &coin,
-    ) {
-        Ok(tx) => tx,
-        Err(e) => {
-            tx_handle.error();
+    if let Err(e) = gas::add_budget(&mut tx, objects, address.into(), &budget_coin) {
+        tx_handle.error();
 
-            return Err(NexusCliError::Any(e));
-        }
-    };
+        return Err(NexusCliError::Any(e));
+    }
 
     tx_handle.success();