+++ cli/src/tool/tool_unregister.rs
@@ -20,25 +20,15 @@ pub(crate) async fn unregister_tool(
     }
 
     // Load CLI configuration.
-    let conf = CliConf::load().await.unwrap_or_else(|_| CliConf::default());
+    let conf = CliConf::load().await.unwrap_or_default();
 
     // Nexus objects must be present in the configuration.
-    let NexusObjects {
-        workflow_pkg_id,
-        tool_registry,
-        ..
-    } = get_nexus_objects(&conf)?;
+    let objects = get_nexus_objects(&conf)?;
 
     // Create wallet context, Sui client and find the active address.
     let mut wallet = create_wallet_context(&conf.sui.wallet_path, conf.sui.net).await?;
     let sui = build_sui_client(&conf.sui).await?;
-
-    let address = match wallet.active_address() {
-        Ok(address) => address,
-        Err(e) => {
-            return Err(NexusCliError::Any(e));
-        }
-    };
+    let address = wallet.active_address().map_err(NexusCliError::Any)?;
 
     // Fetch gas coin object.
     let gas_coin = fetch_gas_coin(&sui, conf.sui.net, address, sui_gas_coin).await?;
@@ -54,20 +44,11 @@ pub(crate) async fn unregister_tool(
 
     let mut tx = sui::ProgrammableTransactionBuilder::new();
 
-    match tool::unregister(
-        &mut tx,
-        &tool_fqn,
-        &owner_cap,
-        tool_registry,
-        *workflow_pkg_id,
-    ) {
-        Ok(tx) => tx,
-        Err(e) => {
-            tx_handle.error();
+    if let Err(e) = tool::unregister(&mut tx, objects, &tool_fqn, &owner_cap) {
+        tx_handle.error();
 
-            return Err(NexusCliError::Any(e));
-        }
-    };
+        return Err(NexusCliError::Any(e));
+    }
 
     tx_handle.success();