+++ sdk/src/transactions/tool.rs
@@ -1,26 +1,23 @@
 use crate::{
     idents::{move_std, primitives, sui_framework, workflow},
     sui,
-    types::ToolMeta,
+    types::{NexusObjects, ToolMeta},
     ToolFqn,
 };
 
 /// PTB template for registering a new Nexus Tool.
 pub fn register_off_chain_for_self(
     tx: &mut sui::ProgrammableTransactionBuilder,
+    objects: &NexusObjects,
     meta: &ToolMeta,
     address: sui::ObjectID,
     collateral_coin: &sui::Coin,
     invocation_cost: u64,
-    tool_registry: &sui::ObjectRef,
-    gas_service: &sui::ObjectRef,
-    workflow_pkg_id: sui::ObjectID,
-    primitives_pkg_id: sui::ObjectID,
 ) -> anyhow::Result<sui::Argument> {
     // `self: &mut ToolRegistry`
     let tool_registry = tx.obj(sui::ObjectArg::SharedObject {
-        id: tool_registry.object_id,
-        initial_shared_version: tool_registry.version,
+        id: objects.tool_registry.object_id,
+        initial_shared_version: objects.tool_registry.version,
         mutable: true,
     })?;
 
@@ -49,7 +46,7 @@ pub fn register_off_chain_for_self(
 
     // `nexus_workflow::tool_registry::register_off_chain_tool()`
     let owner_cap_over_tool = tx.programmable_move_call(
-        workflow_pkg_id,
+        objects.workflow_pkg_id,
         workflow::ToolRegistry::REGISTER_OFF_CHAIN_TOOL
             .module
             .into(),
@@ -69,7 +66,7 @@ pub fn register_off_chain_for_self(
 
     // `nexus_workflow::gas::deescalate()`
     let owner_cap_over_gas = tx.programmable_move_call(
-        workflow_pkg_id,
+        objects.workflow_pkg_id,
         workflow::Gas::DEESCALATE.module.into(),
         workflow::Gas::DEESCALATE.name.into(),
         vec![],
@@ -78,8 +75,8 @@ pub fn register_off_chain_for_self(
 
     // `gas_service: &mut GasService`
     let gas_service = tx.obj(sui::ObjectArg::SharedObject {
-        id: gas_service.object_id,
-        initial_shared_version: gas_service.version,
+        id: objects.gas_service.object_id,
+        initial_shared_version: objects.gas_service.version,
         mutable: true,
     })?;
 
@@ -88,7 +85,7 @@ pub fn register_off_chain_for_self(
 
     // `nexus_workflow::gas::set_single_invocation_cost_mist`
     tx.programmable_move_call(
-        workflow_pkg_id,
+        objects.workflow_pkg_id,
         workflow::Gas::SET_SINGLE_INVOCATION_COST_MIST.module.into(),
         workflow::Gas::SET_SINGLE_INVOCATION_COST_MIST.name.into(),
         vec![],
@@ -103,11 +100,11 @@ pub fn register_off_chain_for_self(
 
     // `CloneableOwnerCap<OverGas>`
     let over_gas_type = sui::MoveTypeTag::Struct(Box::new(sui::MoveStructTag {
-        address: *primitives_pkg_id,
+        address: *objects.primitives_pkg_id,
         module: primitives::OwnerCap::CLONEABLE_OWNER_CAP.module.into(),
         name: primitives::OwnerCap::CLONEABLE_OWNER_CAP.name.into(),
         type_params: vec![sui::MoveTypeTag::Struct(Box::new(sui::MoveStructTag {
-            address: *workflow_pkg_id,
+            address: *objects.workflow_pkg_id,
             module: workflow::Gas::OVER_GAS.module.into(),
             name: workflow::Gas::OVER_GAS.name.into(),
             type_params: vec![],
@@ -116,11 +113,11 @@ pub fn register_off_chain_for_self(
 
     // `CloneableOwnerCap<OverTool>`
     let over_tool_type = sui::MoveTypeTag::Struct(Box::new(sui::MoveStructTag {
-        address: *primitives_pkg_id,
+        address: *objects.primitives_pkg_id,
         module: primitives::OwnerCap::CLONEABLE_OWNER_CAP.module.into(),
         name: primitives::OwnerCap::CLONEABLE_OWNER_CAP.name.into(),
         type_params: vec![sui::MoveTypeTag::Struct(Box::new(sui::MoveStructTag {
-            address: *workflow_pkg_id,
+            address: *objects.workflow_pkg_id,
             module: workflow::ToolRegistry::OVER_TOOL.module.into(),
             name: workflow::ToolRegistry::OVER_TOOL.name.into(),
             type_params: vec![],
@@ -151,18 +148,64 @@ pub fn register_off_chain_for_self(
     ))
 }
 
+/// PTB template for setting the invocation cost of a Nexus Tool.
+pub fn set_invocation_cost(
+    tx: &mut sui::ProgrammableTransactionBuilder,
+    objects: &NexusObjects,
+    tool_fqn: &ToolFqn,
+    owner_cap: &sui::ObjectRef,
+    invocation_cost: u64,
+) -> anyhow::Result<sui::Argument> {
+    // `self: &mut GasService`
+    let gas_service = tx.obj(sui::ObjectArg::SharedObject {
+        id: objects.gas_service.object_id,
+        initial_shared_version: objects.gas_service.version,
+        mutable: true,
+    })?;
+
+    // `tool_registry: &mut ToolRegistry`
+    let tool_registry = tx.obj(sui::ObjectArg::SharedObject {
+        id: objects.tool_registry.object_id,
+        initial_shared_version: objects.tool_registry.version,
+        mutable: true,
+    })?;
+
+    // `owner_cap: &CloneableOwnerCap<OverGas>`
+    let owner_cap = tx.obj(sui::ObjectArg::ImmOrOwnedObject(owner_cap.to_object_ref()))?;
+
+    // `fqn: AsciiString`
+    let fqn = move_std::Ascii::ascii_string_from_str(tx, tool_fqn.to_string())?;
+
+    // `single_invocation_cost_mist: u64`
+    let single_invocation_cost_mist = tx.pure(invocation_cost)?;
+
+    // `nexus_workflow::gas::set_single_invocation_cost_mist`
+    Ok(tx.programmable_move_call(
+        objects.workflow_pkg_id,
+        workflow::Gas::SET_SINGLE_INVOCATION_COST_MIST.module.into(),
+        workflow::Gas::SET_SINGLE_INVOCATION_COST_MIST.name.into(),
+        vec![],
+        vec![
+            gas_service,
+            tool_registry,
+            owner_cap,
+            fqn,
+            single_invocation_cost_mist,
+        ],
+    ))
+}
+
 /// PTB template for unregistering a Nexus Tool.
 pub fn unregister(
     tx: &mut sui::ProgrammableTransactionBuilder,
+    objects: &NexusObjects,
     tool_fqn: &ToolFqn,
     owner_cap: &sui::ObjectRef,
-    tool_registry: &sui::ObjectRef,
-    workflow_pkg_id: sui::ObjectID,
 ) -> anyhow::Result<sui::Argument> {
     // `self: &mut ToolRegistry`
     let tool_registry = tx.obj(sui::ObjectArg::SharedObject {
-        id: tool_registry.object_id,
-        initial_shared_version: tool_registry.version,
+        id: objects.tool_registry.object_id,
+        initial_shared_version: objects.tool_registry.version,
         mutable: true,
     })?;
 
@@ -177,7 +220,7 @@ pub fn unregister(
 
     // `nexus::tool_registry::unregister_tool()`
     Ok(tx.programmable_move_call(
-        workflow_pkg_id,
+        objects.workflow_pkg_id,
         workflow::ToolRegistry::UNREGISTER_TOOL.module.into(),
         workflow::ToolRegistry::UNREGISTER_TOOL.name.into(),
         vec![],
@@ -189,15 +232,14 @@ pub fn unregister(
 /// transferred to the tx sender.
 pub fn claim_collateral_for_self(
     tx: &mut sui::ProgrammableTransactionBuilder,
+    objects: &NexusObjects,
     tool_fqn: &ToolFqn,
     owner_cap: &sui::ObjectRef,
-    tool_registry: &sui::ObjectRef,
-    workflow_pkg_id: sui::ObjectID,
 ) -> anyhow::Result<sui::Argument> {
     // `self: &mut ToolRegistry`
     let tool_registry = tx.obj(sui::ObjectArg::SharedObject {
-        id: tool_registry.object_id,
-        initial_shared_version: tool_registry.version,
+        id: objects.tool_registry.object_id,
+        initial_shared_version: objects.tool_registry.version,
         mutable: true,
     })?;
 
@@ -212,7 +254,7 @@ pub fn claim_collateral_for_self(
 
     // `nexus::tool_registry::claim_collateral_for_tool()`
     Ok(tx.programmable_move_call(
-        workflow_pkg_id,
+        objects.workflow_pkg_id,
         workflow::ToolRegistry::CLAIM_COLLATERAL_FOR_SELF
             .module
             .into(),
@@ -242,25 +284,19 @@ mod tests {
             output_schema: json!({}),
         };
 
+        let objects = sui_mocks::mock_nexus_objects();
         let collateral_coin = sui_mocks::mock_sui_coin(100);
-        let tool_registry = sui_mocks::mock_sui_object_ref();
-        let gas_service = sui_mocks::mock_sui_object_ref();
-        let workflow_pkg_id = sui::ObjectID::random();
-        let primitives_pkd_id = sui::ObjectID::random();
         let address = sui::ObjectID::random();
         let invocation_cost = 1000;
 
         let mut tx = sui::ProgrammableTransactionBuilder::new();
         register_off_chain_for_self(
             &mut tx,
+            &objects,
             &meta,
             address,
             &collateral_coin,
             invocation_cost,
-            &tool_registry,
-            &gas_service,
-            workflow_pkg_id,
-            primitives_pkd_id,
         )
         .expect("Failed to build PTB for registering a tool.");
         let tx = tx.finish();
@@ -269,7 +305,7 @@ mod tests {
             panic!("Expected a command to be a MoveCall to register a tool");
         };
 
-        assert_eq!(call.package, workflow_pkg_id);
+        assert_eq!(call.package, objects.workflow_pkg_id);
 
         assert_eq!(
             call.module,
@@ -290,27 +326,20 @@ mod tests {
 
     #[test]
     fn test_unregister_tool() {
+        let objects = sui_mocks::mock_nexus_objects();
         let tool_fqn = fqn!("xyz.dummy.tool@1");
         let owner_cap = sui_mocks::mock_sui_object_ref();
-        let tool_registry = sui_mocks::mock_sui_object_ref();
-        let workflow_pkg_id = sui::ObjectID::random();
 
         let mut tx = sui::ProgrammableTransactionBuilder::new();
-        unregister(
-            &mut tx,
-            &tool_fqn,
-            &owner_cap,
-            &tool_registry,
-            workflow_pkg_id,
-        )
-        .expect("Failed to build PTB for unregistering a tool.");
+        unregister(&mut tx, &objects, &tool_fqn, &owner_cap)
+            .expect("Failed to build PTB for unregistering a tool.");
         let tx = tx.finish();
 
         let sui::Command::MoveCall(call) = &tx.commands.last().unwrap() else {
             panic!("Expected last command to be a MoveCall to unregister a tool");
         };
 
-        assert_eq!(call.package, workflow_pkg_id);
+        assert_eq!(call.package, objects.workflow_pkg_id);
 
         assert_eq!(
             call.module,
@@ -327,27 +356,20 @@ mod tests {
 
     #[test]
     fn test_claim_collateral_for_self() {
+        let objects = sui_mocks::mock_nexus_objects();
         let tool_fqn = fqn!("xyz.dummy.tool@1");
         let owner_cap = sui_mocks::mock_sui_object_ref();
-        let tool_registry = sui_mocks::mock_sui_object_ref();
-        let workflow_pkg_id = sui::ObjectID::random();
 
         let mut tx = sui::ProgrammableTransactionBuilder::new();
-        claim_collateral_for_self(
-            &mut tx,
-            &tool_fqn,
-            &owner_cap,
-            &tool_registry,
-            workflow_pkg_id,
-        )
-        .expect("Failed to build PTB for claiming collateral for a tool.");
+        claim_collateral_for_self(&mut tx, &objects, &tool_fqn, &owner_cap)
+            .expect("Failed to build PTB for claiming collateral for a tool.");
         let tx = tx.finish();
 
         let sui::Command::MoveCall(call) = &tx.commands.last().unwrap() else {
             panic!("Expected last command to be a MoveCall to claim collateral for a tool");
         };
 
-        assert_eq!(call.package, workflow_pkg_id);
+        assert_eq!(call.package, objects.workflow_pkg_id);
 
         assert_eq!(
             call.module,
@@ -364,5 +386,40 @@ mod tests {
         );
 
         assert_eq!(call.arguments.len(), 4);
+    }
+
+    #[test]
+    fn test_set_invocation_cost() {
+        let tool_fqn = fqn!("xyz.dummy.tool@1");
+        let owner_cap = sui_mocks::mock_sui_object_ref();
+        let objects = sui_mocks::mock_nexus_objects();
+        let invocation_cost = 500;
+
+        let mut tx = sui::ProgrammableTransactionBuilder::new();
+        set_invocation_cost(&mut tx, &objects, &tool_fqn, &owner_cap, invocation_cost)
+            .expect("Failed to build PTB for setting invocation cost.");
+        let tx = tx.finish();
+
+        let sui::Command::MoveCall(call) = &tx.commands.last().unwrap() else {
+            panic!("Expected last command to be a MoveCall to set invocation cost");
+        };
+
+        assert_eq!(call.package, objects.workflow_pkg_id);
+
+        assert_eq!(
+            call.module,
+            workflow::Gas::SET_SINGLE_INVOCATION_COST_MIST
+                .module
+                .to_string(),
+        );
+
+        assert_eq!(
+            call.function,
+            workflow::Gas::SET_SINGLE_INVOCATION_COST_MIST
+                .name
+                .to_string()
+        );
+
+        assert_eq!(call.arguments.len(), 5);
     }
 }