+++ tools/social-twitter/src/list/mod.rs
@@ -1,8 +1,10 @@
 pub(crate) mod add_member;
 pub(crate) mod create_list;
+pub(crate) mod delete_list;
 pub(crate) mod get_list;
 pub(crate) mod get_list_members;
 pub(crate) mod get_list_tweets;
+pub(crate) mod get_user_lists;
 pub(crate) mod models;
 pub(crate) mod remove_member;
 pub(crate) mod update_list;