+++ tools/social-twitter/src/list/models.rs
@@ -1,19 +1,24 @@
 use {
-    crate::tweet::models::{
-        ApiError,
-        Attachments,
-        ContextAnnotation,
-        EditControls,
-        Entities,
-        Geo,
-        NonPublicMetrics,
-        NoteTweet,
-        OrganicMetrics,
-        PromotedMetrics,
-        PublicMetrics,
-        ReferencedTweet,
-        Scopes,
-        Withheld,
+    crate::{
+        error::{TwitterApiError, TwitterError, TwitterErrorKind, TwitterErrorResponse},
+        impl_twitter_response_parser,
+        tweet::models::{
+            ApiError,
+            Attachments,
+            ContextAnnotation,
+            EditControls,
+            Entities,
+            Geo,
+            NonPublicMetrics,
+            NoteTweet,
+            OrganicMetrics,
+            PromotedMetrics,
+            PublicMetrics,
+            ReferencedTweet,
+            Scopes,
+            Withheld,
+        },
+        twitter_client::TwitterApiParsedResponse,
     },
     schemars::JsonSchema,
     serde::{Deserialize, Serialize},
@@ -25,6 +30,18 @@ pub struct ListResponse {
     pub data: Option<ListData>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub errors: Option<Vec<ApiError>>,
+}
+
+#[derive(Debug, Serialize, Deserialize, JsonSchema)]
+pub struct ListsResponse {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub data: Option<Vec<ListData>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub errors: Option<Vec<TwitterApiError>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub includes: Option<Includes>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub meta: Option<Meta>,
 }
 
 #[derive(Debug, Serialize, Deserialize, JsonSchema)]
@@ -273,3 +290,24 @@ pub enum UserField {
     VerifiedType,
     Withheld,
 }
+
+#[derive(Debug, Serialize, Deserialize, JsonSchema)]
+pub struct DeleteListResponse {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub data: Option<DeleteListData>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub errors: Option<Vec<TwitterApiError>>,
+}
+
+#[derive(Debug, Serialize, Deserialize, JsonSchema)]
+pub struct DeleteListData {
+    pub deleted: bool,
+}
+
+impl_twitter_response_parser!(
+    ListsResponse,
+    Vec<ListData>,
+    includes = Includes,
+    meta = Meta
+);
+impl_twitter_response_parser!(DeleteListResponse, DeleteListData);