+++ tools/social-twitter/src/user/mod.rs
@@ -4,4 +4,7 @@
 
 pub(crate) mod get_user_by_id;
 pub(crate) mod get_user_by_username;
+pub(crate) mod get_users_by_username;
+pub(crate) mod get_users_by_id;
 pub(crate) mod models;
+pub(crate) mod unfollow_user;